+++ cypress/integration/vitessce.spec.js
@@ -46,10 +46,10 @@ describe('Vitessce', () => {
 
   it('loads details (responsive)', () => {
     cy.visit('/?dataset=linnarsson-2018');
-    cy.contains('Please wait');
     // https://github.com/cypress-io/cypress/issues/4395
-    // cy.get('.modal-body').should('be.visible');
-    cy.get('.modal-body').should('exist');
+    // cy.get('.modal-dialog').should('be.visible');
+    cy.get('.modal-dialog').should('exist');
+    cy.get('.modal-dialog .please-wait-spinner').should('have.attr', 'title', 'Please wait...');
 
     // Data Set:
     cy.contains('Linnarsson: Spatial organization');